css + menu bar
Please use a separate file for the css of your own page so we can use a shared css for the things we have in all pages, for example, menu bar. I put menu bar into a php file so we can use that when we have back-end implemented. It's easier to handle doing this so we don't have to change the menu bar parts in all files when we want to make changes on it.